requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing-extensions==4.8.0
uritemplate==4.1.1